+++ inspect_wandb/weave/hooks.py
@@ -218,19 +218,12 @@ class WeaveEvaluationHooks(Hooks):
 
         if data.sample.scores is not None:
             for k,v in data.sample.scores.items():
-                score_metadata = (v.metadata or {}) | ({"explanation": v.explanation} if v.explanation is not None else {})
+                score_metadata = (v.metadata or {}) | ({"explanation": v.explanation} if v.explanation is not None else {}) | ({"answer": v.answer} if v.answer is not None else {})
                 with weave.attributes(score_metadata):
-                    if isinstance(v.value, dict):
-                        for k2,v2 in v.value.items():
-                            await sample_score_logger.alog_score(
-                                scorer=f"{k}_{k2}",
-                                score=format_score_types(v2)
-                            )
-                    else:
-                        await sample_score_logger.alog_score(
-                            scorer=k,
-                            score=format_score_types(v.value)
-                        )
+                    await sample_score_logger.alog_score(
+                        scorer=k,
+                        score=format_score_types(v.value, scorer_name=k)
+                    )
 
 
         # Total time