+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Bump minimum Weave version to fix Pydantic validation error on summary aggregation
 - Autopatch based on installed libs, rather than blanket patching
 - Load settings every time enabled is called
+- Remove custom EvaluationLogger override to enable dataset comparison
 
 ### New Contributors
 - [@alex-remedios-aisi](https://github.com/alex-remedios-aisi)