+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Autopatch based on installed libs, rather than blanket patching
 - Load settings every time enabled is called
 - Remove custom EvaluationLogger to enable dataset comparison
+- Bug with out-of-date syntax for importlib utils
 
 ### New Contributors
 - [@alex-remedios-aisi](https://github.com/alex-remedios-aisi)